+++ application/command-composition/src/main/kotlin/org/occurrent/application/composition/command/CompositionExtensions.kt
@@ -16,8 +16,6 @@
 
 package org.occurrent.application.composition.command
 
-import java.util.function.Function
-
 // Sequence Composition
 
 /**
@@ -32,80 +30,31 @@ import java.util.function.Function
  * @param anotherCommand The other command to run after this one.
  */
 
-@JvmName("andThenDifferentTypes")
+/**
+ * Backward compose two functions
+ */
 infix fun <A, B, R> ((A) -> B).andThen(anotherCommand: (B) -> R): (A) -> R = { a ->
     anotherCommand(this(a))
 }
-
-infix fun <T> ((Sequence<T>) -> Sequence<T>).andThen(anotherCommand: (Sequence<T>) -> Sequence<T>): (Sequence<T>) -> Sequence<T> =
-    composeCommands(this, anotherCommand)
 
 /**
  * Compose multiple commands
  */
 fun <T> composeCommands(
-    firstCommand: (Sequence<T>) -> Sequence<T>,
-    secondCommand: (Sequence<T>) -> Sequence<T>,
-    vararg additionalCommands: (Sequence<T>) -> Sequence<T>
-): ((Sequence<T>) -> Sequence<T>) {
+    firstCommand: (T) -> T,
+    secondCommand: (T) -> T,
+    vararg additionalCommands: (T) -> T
+): ((T) -> T) {
     return composeCommands(sequenceOf(firstCommand, secondCommand, *additionalCommands))
 }
 
 /**
  * Compose a sequence of commands
  */
-fun <T> composeCommands(commands: Sequence<(Sequence<T>) -> Sequence<T>>): (Sequence<T>) -> Sequence<T> {
-    val listCommands = commands.map { it.toMutableListCommandFromSequence() }.toList()
-    return { events ->
-        ListCommandComposition.composeCommands(listCommands).apply(events.toList()).asSequence()
+fun <T> composeCommands(commands: Sequence<(T) -> T>): (T) -> T {
+    return { initial ->
+        commands.fold(initial) { acc, cmd ->
+            cmd(acc)
+        }
     }
 }
-
-// List Composition
-/**
- * Compose two commands using infix notation. The resulting command will be executed atomically in the event store.
- * For example:
- * ```kotlin
- * val cmd1 : (List<DomainEvent>) -> List<DomainEvent> = ..
- * val cmd2 : (List<DomainEvent>) -> List<DomainEvent> = ..
- * applicationService.execute("streamId", cmd1 andThen cmd2)
- * ```
- *
- * @param anotherCommand The other command to run after this one.
- */
-@JvmName("andThenList")
-infix fun <T> ((List<T>) -> List<T>).andThen(anotherCommand: (List<T>) -> List<T>): (List<T>) -> List<T> =
-    composeCommands(this, anotherCommand)
-
-/**
- * Compose multiple commands
- */
-@JvmName("commandsListCommands")
-fun <T> composeCommands(
-    firstCommand: (List<T>) -> List<T>,
-    secondCommand: (List<T>) -> List<T>,
-    vararg additionalCommands: (List<T>) -> List<T>
-): ((List<T>) -> List<T>) {
-    return composeCommands(listOf(firstCommand, secondCommand, *additionalCommands))
-}
-
-/**
- * Compose a sequence of commands
- */
-fun <T> composeCommands(commands: List<(List<T>) -> List<T>>): (List<T>) -> List<T> {
-    val listCommands = commands.map { it.toMutableListCommandFromList() }
-    return { events ->
-        ListCommandComposition.composeCommands(listCommands).apply(events.toList())
-    }
-}
-
-// Helpers
-private fun <T> ((Sequence<T>) -> Sequence<T>).toMutableListCommandFromSequence(): Function<MutableList<T>, MutableList<T>> =
-    Function<MutableList<T>, MutableList<T>> { events ->
-        this(events.asSequence()).toMutableList()
-    }
-
-private fun <T> ((List<T>) -> List<T>).toMutableListCommandFromList(): Function<MutableList<T>, MutableList<T>> =
-    Function<MutableList<T>, MutableList<T>> { events ->
-        this(events).toMutableList()
-    }